Changed Turning mechanism to be independent of speed.
Roombas now turn in terms of a turning factor in a certain direction.
Turning factor decreases as the turn in the same direction progresses and resets when another command if taken in or the keys are released.

Higher turning factor means turns are less sharp.

Removed speed cap so roombas travel at max speed.